+++ backend/models/user.py
@@ -1,6 +1,7 @@
-from flask_login import UserMixin
 import psycopg2
-from werkzeug.security import generate_password_hash, check_password_hash
+from flask_login import UserMixin
+from werkzeug.security import check_password_hash, generate_password_hash
+
 
 class User(UserMixin):
     def __init__(self, id, username, email, password_hash):
@@ -38,12 +39,12 @@ class User(UserMixin):
     def check_password(self, password):
         return check_password_hash(self.password_hash, password)
 
-    # @classmethod
-    # def create_user(cls, username, email, password):
-    #     password_hash = generate_password_hash(password)
-    #     conn = psycopg2.connect("dbname=dreamsink user=test password=test host=localhost port=5432")
-    #     cur = conn.cursor()
-    #     cur.execute('INSERT INTO users (username, email, password_hash) VALUES (%s, %s, %s)',
-    #                 (username, email, password_hash))  # 修正
-    #     conn.commit()
-    #     conn.close()
+    @classmethod
+    def create_user(cls, username, email, password):
+        password_hash = generate_password_hash(password)
+        conn = psycopg2.connect("dbname=dreamsink user=test password=test host=localhost port=5432")
+        cur = conn.cursor()
+        cur.execute('INSERT INTO users (username, email, password_hash) VALUES (%s, %s, %s)',
+                    (username, email, password_hash))  # 修正
+        conn.commit()
+        conn.close()
