+++ backend/routes/dream.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, jsonify, request
-from models.dream_models import Dream
+
+from backend.models.dream import Dream
 
 dream_bp = Blueprint('dream', __name__)
