+++ backend/routes/public_dream.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, jsonify
-from models.dream_models import Dream
+
+from backend.models.dream import Dream
 
 public_dream_bp = Blueprint('public_dream', __name__)
